+++ examples/blog/models/blog.ts
@@ -1,5 +1,6 @@
 /// <reference lib="deno.unstable" />
 import { Temporal } from "npm:@js-temporal/polyfill@^0.4.4";
+import { exportJwk, generateCryptoKeyPair, importJwk } from "fedify/httpsig";
 import { hash, verify } from "scrypt";
 import { openKv } from "./kv.ts";
 
@@ -22,24 +23,15 @@ export interface Blog extends BlogBase {
 
 export async function setBlog(blog: BlogInput): Promise<void> {
   const kv = await openKv();
-  const { privateKey, publicKey } = await crypto.subtle.generateKey(
-    {
-      name: "RSASSA-PKCS1-v1_5",
-      modulusLength: 4096,
-      publicExponent: new Uint8Array([0x01, 0x00, 0x01]),
-      hash: "SHA-256",
-    },
-    true,
-    ["sign", "verify"],
-  );
+  const { privateKey, publicKey } = await generateCryptoKeyPair();
   await kv.set(["blog"], {
     handle: blog.handle,
     title: blog.title,
     description: blog.description,
     published: new Date().toISOString(),
     passwordHash: hash(blog.password, undefined, "scrypt"),
-    privateKey: await crypto.subtle.exportKey("jwk", privateKey),
-    publicKey: await crypto.subtle.exportKey("jwk", publicKey),
+    privateKey: await exportJwk(privateKey),
+    publicKey: await exportJwk(publicKey),
   });
 }
 
@@ -56,20 +48,8 @@ export async function getBlog(): Promise<Blog | null> {
   if (entry == null || entry.value == null) return null;
   return {
     ...entry.value,
-    privateKey: await crypto.subtle.importKey(
-      "jwk",
-      entry.value.privateKey,
-      { name: "RSASSA-PKCS1-v1_5", hash: "SHA-256" },
-      true,
-      ["sign"],
-    ),
-    publicKey: await crypto.subtle.importKey(
-      "jwk",
-      entry.value.publicKey,
-      { name: "RSASSA-PKCS1-v1_5", hash: "SHA-256" },
-      true,
-      ["verify"],
-    ),
+    privateKey: await importJwk(entry.value.privateKey, "private"),
+    publicKey: await importJwk(entry.value.publicKey, "public"),
     published: Temporal.Instant.from(entry.value.published),
   };
 }