+++ CHANGES.md
@@ -8,6 +8,12 @@ nav_order: 9
 Fedify changelog
 ================
 
+Version 0.4.0
+-------------
+
+To be released.
+
+
 Version 0.3.0
 -------------